test(various): prevent problem with uninitialized class members
... happening with php 7.4

"PHP Fatal error: Uncaught Error: Call to a member function useNotes() on null"

Reviewed-on: http://gerrit.tine20.com/customers/15345
Tested-by: Philipp Schüle <p.schuele@metaways.de>

+++ tests/tine20/Felamimail/Controller/FolderTest.php
@@ -69,10 +69,12 @@ class Felamimail_Controller_FolderTest extends PHPUnit_Framework_TestCase
         }
         
         // delete all remaining folders from cache of account
-        $folderBackend = new Felamimail_Backend_Folder();
-        $folders = $folderBackend->getMultipleByProperty($this->_account->getId(), 'account_id');
-        foreach ($folders as $folder) {
-            $folderBackend->delete($folder);
+        if ($this->_account) {
+            $folderBackend = new Felamimail_Backend_Folder();
+            $folders = $folderBackend->getMultipleByProperty($this->_account->getId(), 'account_id');
+            foreach ($folders as $folder) {
+                $folderBackend->delete($folder);
+            }
         }
     }
